+++ eslint-rules/index.js
@@ -1,0 +1,7 @@
+'use strict';
+
+module.exports = {
+  rules: {
+    'no-runtime-state-violations': require('./no-runtime-state-violations.js'),
+  },
+};